# src/main/resources/static/openapi/banking-openapi.yaml
openapi: 3.0.3
info:
  title: Brite Banking API
  version: 1.0.0
  description: |
    REST API for the `org.brite.banking` module of the bankingservices Spring Boot application:
    account registration, lookup, deposit/withdrawal, closing (single and bulk), paginated
    account-status search, bank office/ATM location search, bank statements, plus the module's async-notification and
    Resilience4j payment demo endpoints.

    ## Gateway behavior (applies to every endpoint in this document)

    Every path below sits behind two servlet filters (`org.brite.banking.gateway`), which run
    *before* the controllers:

    1. **`BusinessTransactionIdFilter`** generates a UUID per request and returns it in the
       `X-BTID` response header. The same id is stamped on every server log line for that
       request, so quote it when reporting a problem.
    2. **`BankingRateLimitFilter`** requires the customer-id header (default `X-Customer-Id`,
       configurable via `banking.rate-limit.customer-header-name`) and caps each customer at
       `banking.rate-limit.requests-per-day` requests/day (default 1000). A missing/blank
       header is rejected with `400`; an exhausted quota with `429`. Both are **plain-text**
       bodies written by the filter itself, not the JSON/text bodies produced by the
       controllers' exception handler. Every non-rejected response carries
       `X-RateLimit-Limit` and `X-RateLimit-Remaining`.

...
    authentication - the API has no auth layer.

...

...

...

...
servers:
...
security:
  - CustomerId: []
paths:
...
  /v1/api/accounts/withdraw:
    post:
      tags: [Accounts]
      operationId: withdraw
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Withdraw funds
      description: |
        Withdraws `withdrawAmount` from a checking or savings account. The account type is
        derived from the account-number prefix (`CH` -> `CHECKING`, `SV` -> `SAVINGS`) and
        `accountType` in the body **must match it**.

        Checks, in order: account exists; account is not `CLOSED`; sufficient funds; the
        remaining balance must stay at or above the minimum
        (`banking.constraints.checking-minimum-balance` 25.00 / `saving-minimum-balance` 100.00
        by default). A completed withdrawal is recorded in withdrawal history and the
        transaction ledger, and triggers an async email/SMS.

        `withdrawalDate` and `withdrawalStatus` are accepted but ignored - the server stamps
        today's date and `COMPLETED`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WithdrawalRequest'}
            example:
              accountNumber: CH-0000088291
              accountType: CHECKING
              withdrawAmount: 100.75
              firstName: Jane
              lastName: Doe
              street: '111'
              addressLine1: Leafvillage
              city: Leander
              state: TX
              zip: '78717'
      responses:
        '200':
          description: Withdrawal completed; returns the account with its updated balance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): non-positive
            amount, unrecognized prefix, type/prefix mismatch, account closed or suspended,
            insufficient funds, or minimum balance would be violated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                insufficientFunds:
                  value: Insufficient funds in account CH-0000088291
                minimumBalance:
                  value: 'Withdrawal declined: account CH-0000088291 must retain a minimum balance of 25.00'
                closed:
                  value: Account CH-0000010004 is closed and cannot be used for transactions
                suspended:
                  value: Account CH-0000050001 is suspended and cannot be used for transactions until it is reactivated
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/deposit:
    post:
      tags: [Accounts]
      operationId: deposit
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Deposit funds
      description: |
        Deposits `amount` into a checking or savings account. As with withdrawals, the
        account type comes from the account-number prefix and `accountType` must match it.

        `depositType` must be `cash` or `check` (case-insensitive). **Cash deposits are
        capped** at `banking.constraints.maximum-deposit-amount-by-cash` (default 5000.00);
        check deposits are not. A `CLOSED` or `SUSPENDED` account rejects deposits (`400`). Unlike withdrawals,
        deposits do not trigger a notification.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DepositForm'}
            example:
              accountNumber: SV-0000044102
              accountType: SAVINGS
              amount: 250.00
              depositType: check
              firstName: Jane
              lastName: Doe
              street: '111'
              addressLine1: Leafvillage
              city: Leander
              state: TX
              zip: '78717'
      responses:
        '200':
          description: Deposit completed; returns the account with its updated balance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): non-positive
            amount, unrecognized prefix, type/prefix mismatch, invalid deposit type, cash
            limit exceeded, or account closed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                cashLimit:
                  value: Cash deposits cannot exceed 5000.00
                depositType:
                  value: Deposit type must be 'cash' or 'check'
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/close:
    post:
      tags: [Accounts]
      operationId: closeAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Close one account
      description: >-
        Transitions an `ACTIVE` or `SUSPENDED` account to `CLOSED` and stamps `closedDate` with
        today (a suspension is cleared). There is no reopen operation. Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: 'Account closed; returns it with `accountStatus: CLOSED`.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: Account is already closed. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
              example: Account CH-0000010004 is already closed
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/suspend:
    post:
      tags: [Account suspension]
      operationId: suspendAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Suspend an account
      description: |
        `ACTIVE` -> `SUSPENDED`. While suspended the account rejects **every** withdraw and
        deposit (`400`) until it is reactivated, or until `endDateTime` passes and the expiry job
        (`banking.suspension.expiry-job`, every 60s) reactivates it. `endDateTime` omitted = indefinite.

        `startDateTime` defaults to now and must not be in the future (backdating is allowed).
        `endDateTime`, if given, must be after the start and in the future. Both are ISO local
        date-times (`2026-10-01T09:30:00`). Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SuspendAccountRequest'}
            example:
              notes: Suspected fraudulent card activity - under review
              endDateTime: '2026-12-31T17:00:00'
      responses:
        '200':
          description: 'Account suspended; returned with `accountStatus: SUSPENDED` and the suspension fields.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON, e.g. blank notes), or a business rule failed (plain text):
            account closed, already suspended, start in the future, end not after start or in the past.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                alreadySuspended:
                  value: Account CH-0000050001 is already suspended
                closed:
                  value: Account CH-0000010004 is closed and cannot be suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/suspension:
    patch:
      tags: [Account suspension]
      operationId: updateSuspension
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Update a current suspension
      description: |
        Changes the end and/or notes of a suspension; only the fields supplied change, and at
        least one is required. The new `endDateTime` must be in the future and after the stored
        start. The suspension's start can't be changed, and an end can't be removed once set
        (reactivate instead). Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSuspensionRequest'}
            example:
              endDateTime: '2027-01-31T17:00:00'
              notes: Extended pending compliance review
      responses:
        '200':
          description: Updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): nothing to update,
            account isn't suspended, new end in the past or not after the start.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              example: Account CH-0000010001 is not suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/reactivate:
    post:
      tags: [Account suspension]
      operationId: reactivateAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Reactivate a suspended account
      description: |
        `SUSPENDED` -> `ACTIVE`; the suspension flag, dates and notes are cleared and the
        account can transact again. Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: Account is `ACTIVE` again.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: Account isn't suspended. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
              example: Account CH-0000010001 is not suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /v1/api/accounts/{accountNumber}/statement:
    get:
      tags: [Accounts]
      operationId: getStatement
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Generate a bank statement
      description: |
        Returns the account's deposit/withdrawal transactions with `transactionDate` between
        `beginDate` and `endDate` (both inclusive). `endDate` may not precede `beginDate`,
        and the span may not exceed `banking.constraints.max-statement-range-months`
        (default 18). Triggers an async email/SMS notification.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - name: beginDate
          in: query
          required: true
          schema: {type: string, format: date}
        - name: endDate
          in: query
          required: true
          schema: {type: string, format: date}
      responses:
        '200':
          description: The statement.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BankStatement'}
              example:
                accountNumber: CH-0000088291
                beginDate: '2026-08-01'
                endDate: '2026-09-24'
                transactions:
                  - accountNumber: CH-0000088291
                    transactionType: WITHDRAWAL
                    amount: 100.75
                    balanceAfter: 1399.25
                    transactionDate: '2026-09-10'
                    depositType: null
        '400':
          description: >-
            Missing/malformed date parameter (JSON), end date before begin date, or range
            too long (plain text).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                order:
                  value: End date cannot be before begin date
                range:
                  value: Statement date range cannot exceed 18 months
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /bff/v1/portal/home:
    get:
      tags: [Portal (BFF)]
      operationId: portalHome
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Portal home screen bundle
      description: |
        One call for the portal home screen: the newest `ACTIVE` and `SUSPENDED` accounts merged, newest
        first (closed are omitted; at most `banking.portal.home-account-limit`, default 5; no balances; each row has
        `suspended`/`suspendedUntil`) with the active/suspended totals plus branches/ATMs
        sorted by name (at most `banking.portal.nearby-location-limit`, default 5).
        `state` is optional; without it locations are not filtered.
      parameters:
        - name: state
          in: query
          required: false
          description: Two-letter state used to filter nearby locations (case-insensitive).
          schema: {type: string}
          example: TX
      responses:
        '200':
          description: The home bundle.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PortalHomeResponse'}
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/{accountNumber}/overview:
    get:
      tags: [Portal (BFF)]
      operationId: portalAccountOverview
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Account detail with balance and recent activity
      description: |
        Balance plus the account's transactions from the last `days` days, newest first, capped
        at `banking.portal.activity-limit` (default 20). Read-only: unlike the statement
        endpoint it sends no email/SMS.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - name: days
          in: query
          required: false
          description: Activity window in days, 1 to `banking.portal.max-activity-days` (default 90). Defaults to 30.
          schema: {type: integer, minimum: 1, maximum: 90, default: 30}
      responses:
        '200':
          description: The account overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          $ref: '#/components/responses/BadRequestText'
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/withdraw:
    post:
      tags: [Portal (BFF)]
      operationId: portalWithdraw
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Withdraw and get the refreshed account overview
      description: |
        Same body, rules and side effects as `POST /v1/api/accounts/withdraw`, but the response is the
        account overview (balance, suspension state, recent activity) so the portal can redraw in one call.
        A `CLOSED` or `SUSPENDED` account is rejected with a plain-text `400`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WithdrawalRequest'}
      responses:
        '200':
          description: Withdrawal completed; the refreshed overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          description: >-
            Validation failure (JSON) or a business rule failed (plain text), including
            "is suspended and cannot be used for transactions until it is reactivated".
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/deposit:
    post:
      tags: [Portal (BFF)]
      operationId: portalDeposit
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Deposit and get the refreshed account overview
      description: |
        Same body, rules and side effects as `POST /v1/api/accounts/deposit`, but the response is the
        account overview. A `CLOSED` or `SUSPENDED` account is rejected with a plain-text `400`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DepositForm'}
      responses:
        '200':
          description: Deposit completed; the refreshed overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          description: Validation failure (JSON) or a business rule failed (plain text).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/{accountNumber}/suspend:
    post:
      tags: [Portal (BFF)]
      operationId: portalSuspendAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Suspend an account and get the refreshed overview
      description: |
        Same body and rules as `POST /v1/api/accounts/{accountNumber}/suspend`; returns the account
        overview with `suspended: true` and `suspendedUntil` (null = indefinite). Suspension notes are not
        part of the overview.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SuspendAccountRequest'}
            example:
              notes: Suspected fraudulent card activity - under review
              endDateTime: '2026-12-31T17:00:00'
      responses:
        '200':
          description: Account suspended; the refreshed overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          description: Validation failure (JSON) or a business rule failed (plain text), for example closed, already suspended, or an invalid start/end.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/{accountNumber}/suspension:
    patch:
      tags: [Portal (BFF)]
      operationId: portalUpdateSuspension
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Update a suspension and get the refreshed overview
      description: Same body and rules as `PATCH /v1/api/accounts/{accountNumber}/suspension`; returns the account overview.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSuspensionRequest'}
      responses:
        '200':
          description: Suspension updated; the refreshed overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          description: Validation failure (JSON) or a business rule failed (plain text), for example the account isn't suspended.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /bff/v1/portal/accounts/{accountNumber}/reactivate:
    post:
      tags: [Portal (BFF)]
      operationId: portalReactivateAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Reactivate a suspended account and get the refreshed overview
      description: 'Same rules as `POST /v1/api/accounts/{accountNumber}/reactivate`; returns the account overview (`ACTIVE`, `suspended: false`).'
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: Account is active again; the refreshed overview.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountOverviewResponse'}
        '400':
          $ref: '#/components/responses/BadRequestText'
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: |
        Customer identifier used solely for per-customer daily rate limiting - **not**
        authentication; any non-blank value is accepted. The header name is configurable
        (`banking.rate-limit.customer-header-name`) and the check can be turned off with
        `banking.rate-limit.enabled: false`.

    CustomerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The customer access token returned by `POST /v1/api/customers/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. Required on every `/v1/api/accounts/**` and `/bff/v1/portal/**` operation except
        the two that create a customer (`POST /v1/api/accounts/newaccount`, `POST /bff/v1/portal/accounts/open`). It proves who
        logged in; the customer's login is re-checked on every request, and the customer can only reach their own accounts.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The employee access token returned by `POST /v1/api/staff/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. It proves who logged in; permissions and status are re-checked from the
        database on every request, so a suspended login stops working immediately.
...